fix: apply review findings on the single-query group titles page
Ten findings from a high-effort review, each independently verified before
being acted on.

**The one behavior regression.** The `total == 0` early return fired before
pagination normalization, so a group whose titles had all been deleted from the
catalogue returned `{"Page":0,"Size":0,...,"Content":[]}` where the base branch
returned `{"Page":1,"Size":20,...,"Content":null}` — a client computing
`ceil(total/size)` divided by zero, and `null` vs `[]` is an observable
contract (CONVENTIONS §5).

The three empty-page shapes were established empirically, by running a probe
against the base commit in a throwaway worktree rather than by reasoning: `[]`
with raw size/page iff the group holds no entry matching the filters, `null`
with normalized size/page otherwise. An orphan's title-side columns are NULL,
so it can never satisfy an explicit type filter — which is why the base
branch excluded it there and included it everywhere else. `GroupHasTitleEntries`
reproduces that split and is consulted only on the already-empty path.

**Two latent hazards introduced by the refactor.** The new offset guard divided
by the caller-supplied size, so `size == 0` panicked with an integer
divide-by-zero — and this repo has no panic recovery, so that takes the process
down rather than one request. And `int32(size)` wrapped negative if
MAX_PAGE_SIZE were configured above 2^31, making Postgres reject every query.
Both stores now return an empty page for a non-positive size, and LIMIT/OFFSET
are bigint in the queries so sqlc generates int64 — the narrowing is gone
rather than guarded.

**The test gap.** `tiedSortKeys` still excluded `watched`/`watchedAt`/`addedAt`
even though this branch moved exactly those three onto the paged path the tied
sort tests exist to guard. They are covered now, and the fixture had to be
taught to tie on those columns: a control run proves `addedAt` passes vacuously
without it, so the helper is load-bearing rather than decorative.

**Cleanups.** The duplicated offset guard and the duplicated pagination
normalization are each one implementation now, the series-type set has a
canonical spelling, the season-rows query is skipped when a page holds no
series, and the redundant second `GroupExists` is gone — the handler guard is
authoritative because it owns the 404 message the tests pin, and the handler
gained the standard ErrorMap branch it was missing.

`TitleType` keeps reading `titles.type`: `group_titles.title_type` is written
once at add-to-group time and hard-codes "movie" for every title. Correct as
changed, but it was undocumented, so it is now stated in code and pinned by a
test.

All new tests are mutation-checked. The size-guard tests assert through
NotPanics so removing the guard reports a legible failure on the offending
subtest instead of aborting the test binary.

+++ sql/queries/titles.sql
@@ -32,6 +32,11 @@ SELECT count(*) FROM titles;
 -- pre-normalized (unknown keys -> ''). No NULLS FIRST/LAST anywhere: Postgres'
 -- defaults keep deciding where NULL sort values land (added_at/updated_at are
 -- the only nullable keys), exactly as the hand-written version behaved.
+--
+-- page_size/page_offset are cast to bigint so sqlc generates int64 params:
+-- Go's page/size are plain ints, and narrowing them to int32 could wrap
+-- negative (MAX_PAGE_SIZE is env-configurable without an upper bound), which
+-- Postgres rejects with "LIMIT/OFFSET must not be negative".
 SELECT id, primary_title, type, start_year, rating_aggregate, vote_count, added_at, updated_at, metadata
 FROM titles
 ORDER BY
@@ -50,4 +55,4 @@ ORDER BY
     CASE WHEN sqlc.arg('order_by')::text = 'updatedAt'  AND NOT sqlc.arg('descending')::bool THEN updated_at END ASC,
     CASE WHEN sqlc.arg('order_by')::text = 'updatedAt'  AND sqlc.arg('descending')::bool     THEN updated_at END DESC,
     id ASC
-LIMIT sqlc.arg('page_size') OFFSET sqlc.arg('page_offset');
+LIMIT sqlc.arg('page_size')::bigint OFFSET sqlc.arg('page_offset')::bigint;
